Group Yospace types into dedicated packages
Organize the integration source into subpackages mirroring the config
convention:
- advertising: Ad, AdBreak, AdBreakPosition, AdData, CompanionAd,
  AdTimeline, BitmovinTruexAdRenderer
- events: emitter, dispatchers, YospacePlayerEvent and related listeners
- policy: player policy interface, default, and Yospace policy handler
- config: YospaceAssetType, YospaceLiveInitializationType
- deficiency: YospaceErrorCode and YospaceWarningCode (split into
  separate files)

Package moves span shared consumers (BitmovinYospacePlayer, the sample
app, tests), so they are committed together to keep every commit
building.
