+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/advertising/CompanionAd.kt
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integration.yospace
+package com.bitmovin.player.integration.yospace.advertising
 
 enum class CompanionAdType {
     HTML, // dynamic, web view
