+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/YospaceAssetType.kt
@@ -1,11 +1,11 @@
-package com.bitmovin.player.integration.yospace
+package com.bitmovin.player.integration.yospace.config
 
 enum class YospaceAssetType {
     @Deprecated(
         message = "Use LINEAR_START_OVER for DVR live playback.",
         replaceWith = ReplaceWith(
             "YospaceAssetType.LINEAR_START_OVER",
-            "com.bitmovin.player.integration.yospace.YospaceAssetType"
+            "com.bitmovin.player.integration.yospace.config.YospaceAssetType"
         )
     )
     LINEAR,
