+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/YospaceLiveInitializationType.kt
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integration.yospace
+package com.bitmovin.player.integration.yospace.config
 
 enum class YospaceLiveInitializationType {
     PROXY,
