+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/events/YospacePlayerEventListener.kt
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integration.yospace
+package com.bitmovin.player.integration.yospace.events
 
 /**
  * Java-friendly listener for [YospacePlayerEvent]s. Use with the `Class<E>` overloads of
