+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/policy/BitmovinYospacePlayerPolicy.kt
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integration.yospace
+package com.bitmovin.player.integration.yospace.policy
 
 interface BitmovinYospacePlayerPolicy {
